Update MCreator to 2025.2 and remove textfield sync plugin

+++ src/main/java/tannyjung/tanshugetrees/procedures/SaplingBlockRightClickProcedure.java
@@ -49,7 +49,7 @@ public class SaplingBlockRightClickProcedure {
 		}.getValue(world, BlockPos.containing(x, y, z), "sapling_start")) == false) {
 			if ((entity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY).getItem() == Items.BONE_MEAL) {
 				if ((entity instanceof Player _plr ? _plr.getAbilities().instabuild : false) == false) {
-					(entity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY).setCount((int) ((entity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY).getCount() - 1));
+					(entity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY).setCount((entity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY).getCount() - 1);
 					if (!world.isClientSide()) {
 						BlockPos _bp = BlockPos.containing(x, y, z);
 						BlockEntity _blockEntity = world.getBlockEntity(_bp);
